# docs/swagger.yaml
basePath: /
definitions:
...
  fortyfour-backend_internal_dto.CreatePICRequest:
    properties:
      email:
        type: string
      id_perusahaan:
        type: string
      nama:
        type: string
      telepon:
        type: string
    type: object
  fortyfour-backend_internal_dto.CreatePerusahaanRequest:
    properties:
      alamat:
        type: string
      email:
        type: string
      id_sub_sektor:
        type: string
      nama_perusahaan:
        type: string
      photo:
        type: string
      telepon:
        type: string
      website:
        type: string
    type: object
...
  fortyfour-backend_internal_dto.ErrorResponse:
    properties:
      message:
        example: invalid credentials
        type: string
    type: object
...
  fortyfour-backend_internal_dto.MessageResponse:
    properties:
      message:
        example: Logged out successfully
        type: string
    type: object
...
  fortyfour-backend_internal_dto.PICResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      nama:
        type: string
      perusahaan:
        $ref: '#/definitions/fortyfour-backend_internal_dto.PerusahaanInPIC'
      telepon:
        type: string
      updated_at:
        type: string
    type: object
  fortyfour-backend_internal_dto.PerusahaanInPIC:
    properties:
      id:
        type: string
      nama_perusahaan:
        type: string
    type: object
...
  fortyfour-backend_internal_dto.PerusahaanResponse:
    properties:
      alamat:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      nama_perusahaan:
        type: string
      photo:
        type: string
      sub_sektor:
        $ref: '#/definitions/fortyfour-backend_internal_dto.SubSektorResponse'
      telepon:
        type: string
      updated_at:
        type: string
      website:
        type: string
    type: object
...
  fortyfour-backend_internal_dto.SubSektorResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      id_sektor:
        type: string
      nama_sektor:
        type: string
      nama_sub_sektor:
        type: string
      updated_at:
        type: string
    type: object
...
  fortyfour-backend_internal_dto.UpdatePICRequest:
    properties:
      email:
        type: string
      id_perusahaan:
        type: string
      nama:
        type: string
      telepon:
        type: string
    type: object
  fortyfour-backend_internal_dto.UpdatePerusahaanRequest:
    properties:
      alamat:
        type: string
      email:
        type: string
      id_sub_sektor:
        type: string
      nama_perusahaan:
        type: string
      photo:
        type: string
      telepon:
        type: string
      website:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for Fortyfour Backend - main auth and management
    service.
  title: Fortyfour Backend API
  version: "1.0"
paths:
...
  /api/events:
    get:
      description: Establishes an SSE connection for real-time updates
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Server-Sent Events connection
      tags:
      - SSE
  /api/events/stats:
    get:
      description: Returns SSE connection statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get SSE statistics
      tags:
      - SSE
...
  /api/perusahaan:
    get:
      description: Mengambil seluruh data perusahaan (AUTHENTICATED - full data)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fortyfour-backend_internal_dto.PerusahaanResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List semua perusahaan
      tags:
      - Perusahaan
    post:
      consumes:
      - application/json
      description: Membuat record perusahaan
      parameters:
      - description: Data perusahaan
        in: body
        name: perusahaan
        required: true
        schema:
          $ref: '#/definitions/fortyfour-backend_internal_dto.CreatePerusahaanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.PerusahaanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah perusahaan baru
      tags:
      - Perusahaan
  /api/perusahaan/{id}:
    delete:
      description: Menghapus data perusahaan berdasarkan ID
      parameters:
      - description: Perusahaan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus perusahaan
      tags:
      - Perusahaan
    get:
      description: Mengambil satu data perusahaan
      parameters:
      - description: Perusahaan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.PerusahaanResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil perusahaan berdasarkan ID
      tags:
      - Perusahaan
    put:
      consumes:
      - application/json
      description: Mengubah data perusahaan berdasarkan ID
      parameters:
      - description: Perusahaan ID
        in: path
        name: id
        required: true
        type: string
      - description: Data update
        in: body
        name: perusahaan
        required: true
        schema:
          $ref: '#/definitions/fortyfour-backend_internal_dto.UpdatePerusahaanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.PerusahaanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update perusahaan
      tags:
      - Perusahaan
...
  /api/pic:
    get:
      description: Mengambil seluruh data pic perusahaan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fortyfour-backend_internal_dto.PICResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List semua pic perusahaan
      tags:
      - PIC
    post:
      consumes:
      - application/json
      description: Membuat record pic perusahaan
      parameters:
      - description: Data pic perusahaan
        in: body
        name: pic
        required: true
        schema:
          $ref: '#/definitions/fortyfour-backend_internal_dto.CreatePICRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.PICResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah pic perusahaan baru
      tags:
      - PIC
  /api/pic/{id}:
    delete:
      description: Menghapus data pic perusahaan berdasarkan ID
      parameters:
      - description: PIC ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus pic perusahaan
      tags:
      - PIC
    get:
      description: Mengambil satu data pic perusahaan
      parameters:
      - description: PIC ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.PICResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil pic perusahaan berdasarkan ID
      tags:
      - PIC
    put:
      consumes:
      - application/json
      description: Mengubah data pic perusahaan berdasarkan ID
      parameters:
      - description: PIC ID
        in: path
        name: id
        required: true
        type: string
      - description: Data update
        in: body
        name: pic
        required: true
        schema:
          $ref: '#/definitions/fortyfour-backend_internal_dto.UpdatePICRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.PICResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update pic perusahaan
      tags:
      - PIC
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"